refactor(conversations,webui,tools): Update test mocks, log levels, and error rendering
