+++ pkg/conversations/json_store_test.go
@@ -9,6 +9,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/jingkaihe/kodelet/pkg/types/tools"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -311,4 +312,72 @@ func TestGetMostRecentConversationID(t *testing.T) {
 		// and should be returned as the most recent conversation
 		assert.Equal(t, "created-second", conversations[0].ID, "Should return conversation with most recent updated_at timestamp")
 	})
+}
+
+func TestJSONStore_StructuredToolResults(t *testing.T) {
+	tempDir, cleanup := setupTestDir(t)
+	defer cleanup()
+
+	// Create a store
+	store, err := NewJSONConversationStore(context.Background(), tempDir)
+	require.NoError(t, err)
+	defer store.Close()
+
+	// Create a conversation record with structured tool results
+	record := NewConversationRecord("test-conversation")
+	record.ToolResults = map[string]tools.StructuredToolResult{
+		"call_1": {
+			ToolName:  "file_read",
+			Success:   true,
+			Timestamp: time.Now(),
+			Metadata: tools.FileReadMetadata{
+				FilePath:  "/test/file.go",
+				Lines:     []string{"package main", "func main() {}"},
+				Language:  "go",
+				Truncated: false,
+			},
+		},
+		"call_2": {
+			ToolName:  "bash",
+			Success:   true,
+			Timestamp: time.Now(),
+			Metadata: tools.BashMetadata{
+				Command:       "go test ./...",
+				ExitCode:      0,
+				Output:        "ok\tgithub.com/test\t0.005s",
+				ExecutionTime: 5 * time.Second,
+				WorkingDir:    "/test",
+			},
+		},
+	}
+
+	// Save the record
+	err = store.Save(record)
+	require.NoError(t, err)
+
+	// Load it back
+	loaded, err := store.Load(record.ID)
+	require.NoError(t, err)
+
+	// Verify the tool results were preserved
+	assert.Equal(t, len(record.ToolResults), len(loaded.ToolResults), "Tool results count should match")
+
+	// Check specific results
+	for key, original := range record.ToolResults {
+		loaded, exists := loaded.ToolResults[key]
+		assert.True(t, exists, "Tool result should exist for key %s", key)
+
+		assert.Equal(t, original.ToolName, loaded.ToolName, "Tool name should match for %s", key)
+		assert.Equal(t, original.Success, loaded.Success, "Success should match for %s", key)
+		assert.NotNil(t, loaded.Metadata, "Metadata should not be nil for %s", key)
+		assert.Equal(t, original.Metadata.ToolType(), loaded.Metadata.ToolType(), "Metadata type should match for %s", key)
+	}
+
+	// Verify the actual JSON file exists and check its structure
+	jsonPath := filepath.Join(tempDir, record.ID+".json")
+	data, err := os.ReadFile(jsonPath)
+	require.NoError(t, err)
+
+	// The JSON should contain metadataType fields
+	assert.Contains(t, string(data), `"metadataType"`, "JSON should contain metadataType field")
 }
